+++ benches/conversions.rs
@@ -1,6 +1,6 @@
-use dasp_sample::FromSample;
+use dasp_sample::{Duplex, Sample};
 use divan::Bencher;
-use rodio::{decoder::DecoderSample, Source};
+use rodio::conversions::SampleTypeConverter;
 
 mod shared;
 
@@ -9,12 +9,15 @@ fn main() {
 }
 
 #[divan::bench(types = [i16, u16, f32])]
-fn from_sample_to<T: rodio::Sample + FromSample<DecoderSample>>(bencher: Bencher) {
+fn from_sample<S: Duplex<f32>>(bencher: Bencher) {
     bencher
-        .with_inputs(|| shared::music_wav())
+        .with_inputs(|| {
+            shared::music_wav()
+                .map(|s| s.to_sample::<S>())
+                .collect::<Vec<_>>()
+                .into_iter()
+        })
         .bench_values(|source| {
-            source
-                .convert_samples::<T>()
-                .for_each(divan::black_box_drop)
+            SampleTypeConverter::<_, rodio::Sample>::new(source).for_each(divan::black_box_drop)
         })
 }