+++ src/sink.rs
@@ -10,14 +10,14 @@ use std::sync::mpsc::{Receiver, Sender};
 
 use crate::mixer::Mixer;
 use crate::source::SeekError;
-use crate::{queue, source::Done, Sample, Source};
+use crate::{queue, source::Done, Source};
 
 /// Handle to a device that outputs sounds.
 ///
 /// Dropping the `Sink` stops all its sounds. You can use `detach` if you want the sounds to continue
 /// playing.
 pub struct Sink {
-    queue_tx: Arc<queue::SourcesQueueInput<f32>>,
+    queue_tx: Arc<queue::SourcesQueueInput>,
     sleep_until_end: Mutex<Option<Receiver<()>>>,
 
     controls: Arc<Controls>,
@@ -50,7 +50,6 @@ impl SeekOrder {
     fn attempt<S>(self, maybe_seekable: &mut S)
     where
         S: Source,
-        S::Item: Sample + Send,
     {
         let res = maybe_seekable.try_seek(self.pos);
         let _ignore_receiver_dropped = self.feedback.send(res);
@@ -70,7 +69,7 @@ struct Controls {
 impl Sink {
     /// Builds a new `Sink`, beginning playback on a stream.
     #[inline]
-    pub fn connect_new(mixer: &Mixer<f32>) -> Sink {
+    pub fn connect_new(mixer: &Mixer) -> Sink {
         let (sink, source) = Sink::new();
         mixer.add(source);
         sink
@@ -78,7 +77,7 @@ impl Sink {
 
     /// Builds a new `Sink`.
     #[inline]
-    pub fn new() -> (Sink, queue::SourcesQueueOutput<f32>) {
+    pub fn new() -> (Sink, queue::SourcesQueueOutput) {
         let (queue_tx, queue_rx) = queue::queue(true);
 
         let sink = Sink {
@@ -105,7 +104,6 @@ impl Sink {
     where
         S: Source + Send + 'static,
         f32: FromSample<S::Item>,
-        S::Item: Sample + Send,
     {
         // Wait for the queue to flush then resume stopped playback
         if self.controls.stopped.load(Ordering::SeqCst) {
@@ -155,8 +153,7 @@ impl Sink {
                     seek.attempt(amp)
                 }
                 start_played.store(true, Ordering::SeqCst);
-            })
-            .convert_samples();
+            });
         self.sound_count.fetch_add(1, Ordering::Relaxed);
         let source = Done::new(source, self.sound_count.clone());
         *self.sleep_until_end.lock().unwrap() = Some(self.queue_tx.append_with_signal(source));
@@ -373,31 +370,36 @@ mod tests {
 
     #[test]
     fn test_pause_and_stop() {
-        let (sink, mut queue_rx) = Sink::new();
+        let (sink, mut source) = Sink::new();
 
-        // assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(source.next(), Some(0.0));
+        // TODO (review) How did this test passed before? I might have broken something but
+        //      silence source should come first as next source is only polled while previous ends.
+        //      Respective test in Queue seem to be ignored (see queue::test::no_delay_when_added()
+        //      at src/queue.rs:293).
+        let mut source = source.skip_while(|x| *x == 0.0);
 
-        let v = vec![10i16, -10, 20, -20, 30, -30];
+        let v = vec![10.0, -10.0, 20.0, -20.0, 30.0, -30.0];
 
         // Low rate to ensure immediate control.
         sink.append(SamplesBuffer::new(1, 1, v.clone()));
-        let mut src = SamplesBuffer::new(1, 1, v).convert_samples();
+        let mut reference_src = SamplesBuffer::new(1, 1, v);
 
-        assert_eq!(queue_rx.next(), src.next());
-        assert_eq!(queue_rx.next(), src.next());
+        assert_eq!(source.next(), reference_src.next());
+        assert_eq!(source.next(), reference_src.next());
 
         sink.pause();
 
-        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(source.next(), Some(0.0));
 
         sink.play();
 
-        assert_eq!(queue_rx.next(), src.next());
-        assert_eq!(queue_rx.next(), src.next());
+        assert_eq!(source.next(), reference_src.next());
+        assert_eq!(source.next(), reference_src.next());
 
         sink.stop();
 
-        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(source.next(), Some(0.0));
 
         assert_eq!(sink.empty(), true);
     }
@@ -406,10 +408,10 @@ mod tests {
     fn test_stop_and_start() {
         let (sink, mut queue_rx) = Sink::new();
 
-        let v = vec![10i16, -10, 20, -20, 30, -30];
+        let v = vec![10.0, -10.0, 20.0, -20.0, 30.0, -30.0];
 
         sink.append(SamplesBuffer::new(1, 1, v.clone()));
-        let mut src = SamplesBuffer::new(1, 1, v.clone()).convert_samples();
+        let mut src = SamplesBuffer::new(1, 1, v.clone());
 
         assert_eq!(queue_rx.next(), src.next());
         assert_eq!(queue_rx.next(), src.next());
@@ -419,7 +421,7 @@ mod tests {
         assert!(sink.controls.stopped.load(Ordering::SeqCst));
         assert_eq!(queue_rx.next(), Some(0.0));
 
-        src = SamplesBuffer::new(1, 1, v.clone()).convert_samples();
+        src = SamplesBuffer::new(1, 1, v.clone());
         sink.append(SamplesBuffer::new(1, 1, v));
 
         assert!(!sink.controls.stopped.load(Ordering::SeqCst));
@@ -434,11 +436,11 @@ mod tests {
     fn test_volume() {
         let (sink, mut queue_rx) = Sink::new();
 
-        let v = vec![10i16, -10, 20, -20, 30, -30];
+        let v = vec![10.0, -10.0, 20.0, -20.0, 30.0, -30.0];
 
         // High rate to avoid immediate control.
         sink.append(SamplesBuffer::new(2, 44100, v.clone()));
-        let src = SamplesBuffer::new(2, 44100, v.clone()).convert_samples();
+        let src = SamplesBuffer::new(2, 44100, v.clone());
 
         let mut src = src.amplify(0.5);
         sink.set_volume(0.5);