+++ src/source/empty_callback.rs
@@ -1,46 +1,36 @@
-use std::marker::PhantomData;
 use std::time::Duration;
 
 use super::SeekError;
 use crate::common::{ChannelCount, SampleRate};
 use crate::{Sample, Source};
 
-/// An empty source which executes a callback function
-pub struct EmptyCallback<S> {
-    #[allow(missing_docs)] // See: https://github.com/RustAudio/rodio/issues/615
-    pub phantom_data: PhantomData<S>,
-    #[allow(missing_docs)] // See: https://github.com/RustAudio/rodio/issues/615
-    pub callback: Box<dyn Send + Fn()>,
+/// An empty source that executes a callback function
+pub struct EmptyCallback {
+    callback: Box<dyn Send + Fn()>,
 }
 
-impl<S> EmptyCallback<S> {
+impl EmptyCallback {
     #[inline]
-    /// Create an empty source which executes a callback function.
+    /// Create an empty source that executes a callback function.
     /// Example use-case:
     ///
     /// Detect and do something when the source before this one has ended.
-    pub fn new(callback: Box<dyn Send + Fn()>) -> EmptyCallback<S> {
-        EmptyCallback {
-            phantom_data: PhantomData,
-            callback,
-        }
+    pub fn new(callback: Box<dyn Send + Fn()>) -> EmptyCallback {
+        EmptyCallback { callback }
     }
 }
 
-impl<S> Iterator for EmptyCallback<S> {
-    type Item = S;
+impl Iterator for EmptyCallback {
+    type Item = Sample;
 
     #[inline]
-    fn next(&mut self) -> Option<S> {
+    fn next(&mut self) -> Option<Self::Item> {
         (self.callback)();
         None
     }
 }
 
-impl<S> Source for EmptyCallback<S>
-where
-    S: Sample,
-{
+impl Source for EmptyCallback {
     #[inline]
     fn current_span_len(&self) -> Option<usize> {
         None