+++ src/source/periodic.rs
@@ -2,13 +2,12 @@ use std::time::Duration;
 
 use super::SeekError;
 use crate::common::{ChannelCount, SampleRate};
-use crate::{Sample, Source};
+use crate::Source;
 
 /// Internal function that builds a `PeriodicAccess` object.
 pub fn periodic<I, F>(source: I, period: Duration, modifier: F) -> PeriodicAccess<I, F>
 where
     I: Source,
-    I::Item: Sample,
 {
     // TODO: handle the fact that the samples rate can change
     // TODO: generally, just wrong
@@ -47,7 +46,7 @@ pub struct PeriodicAccess<I, F> {
 impl<I, F> PeriodicAccess<I, F>
 where
     I: Source,
-    I::Item: Sample,
+
     F: FnMut(&mut I),
 {
     /// Returns a reference to the inner source.
@@ -72,7 +71,7 @@ where
 impl<I, F> Iterator for PeriodicAccess<I, F>
 where
     I: Source,
-    I::Item: Sample,
+
     F: FnMut(&mut I),
 {
     type Item = I::Item;
@@ -97,7 +96,7 @@ where
 impl<I, F> Source for PeriodicAccess<I, F>
 where
     I: Source,
-    I::Item: Sample,
+
     F: FnMut(&mut I),
 {
     #[inline]
@@ -137,7 +136,7 @@ mod tests {
     #[test]
     fn stereo_access() {
         // Stereo, 1Hz audio buffer
-        let inner = SamplesBuffer::new(2, 1, vec![10i16, -10, 10, -10, 20, -20]);
+        let inner = SamplesBuffer::new(2, 1, vec![10.0, -10.0, 10.0, -10.0, 20.0, -20.0]);
 
         let cnt = RefCell::new(0);
 
@@ -147,25 +146,25 @@ mod tests {
 
         assert_eq!(*cnt.borrow(), 0);
         // Always called on first access!
-        assert_eq!(source.next(), Some(10));
+        assert_eq!(source.next(), Some(10.0));
         assert_eq!(*cnt.borrow(), 1);
         // Called every 1 second afterwards
-        assert_eq!(source.next(), Some(-10));
+        assert_eq!(source.next(), Some(-10.0));
         assert_eq!(*cnt.borrow(), 1);
-        assert_eq!(source.next(), Some(10));
+        assert_eq!(source.next(), Some(10.0));
         assert_eq!(*cnt.borrow(), 2);
-        assert_eq!(source.next(), Some(-10));
+        assert_eq!(source.next(), Some(-10.0));
         assert_eq!(*cnt.borrow(), 2);
-        assert_eq!(source.next(), Some(20));
+        assert_eq!(source.next(), Some(20.0));
         assert_eq!(*cnt.borrow(), 3);
-        assert_eq!(source.next(), Some(-20));
+        assert_eq!(source.next(), Some(-20.0));
         assert_eq!(*cnt.borrow(), 3);
     }
 
     #[test]
     fn fast_access_overflow() {
         // 1hz is lower than 0.5 samples per 5ms
-        let inner = SamplesBuffer::new(1, 1, vec![10i16, -10, 10, -10, 20, -20]);
+        let inner = SamplesBuffer::new(1, 1, vec![10.0, -10.0, 10.0, -10.0, 20.0, -20.0]);
         let mut source = inner.periodic_access(Duration::from_millis(5), |_src| {});
 
         source.next();